requirements.txt
yt-dlp>=2024.1.0
openai-whisper>=20231117
requests>=2.28.0
